moving things around and breaking stuff, hopefully not too bad
