+++ wizard/components/metadata.py
@@ -1,5 +1,6 @@
-from st_state import Metadata
-from st_views.st_utils import needed_field
+from state import Metadata
+from state import CurrentStep
+from utils import needed_field
 import streamlit as st
 
 # List from https://www.kaggle.com/discussions/general/116302.
@@ -22,6 +23,8 @@ licenses = [
 
 
 def render_metadata():
+    if st.session_state[CurrentStep] == "start":
+        "Start Entering Metadata about your files:"
     name = st.text_input(
         label=needed_field("Name"),
         value=st.session_state[Metadata].name,
