+++ wizard/views/jsonld.py
@@ -1,7 +1,7 @@
 import pandas as pd
-from st_state import Files
-from st_state import Metadata
-from st_state import RecordSets
+from state import Files
+from state import Metadata
+from state import RecordSets
 import streamlit as st
 
 import mlcroissant as mlc
